Initial version of the framework
  The framework is capable of loading a website's content as JSON and
rendering them in HTML.
The content in JSON can be text or multimedia(images, audio, video).
It can also load external JS and CSS plugins. (More in docs.)

The code resides in js/framework.js. The markup used(assumed) in code is
in index.html
Rest are boilerplate code or library files.